input/uinput: support abs resolution
Input resolution has been added in kernel v2.6.31 with commit
ec20a022aa24fc63. However, uinput was never updated since, meaning that
devices cannot be emulated in full. User-space processes that rely on the
resolution to be accurate need this.

Resolution is added in struct uinput_user_dev in a way it won't break
current programs relying on the old implementation.

Bump the UINPUT_VERSION number to 4 to aid detection of this in user-space.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>